+++ src/main/java/com/spm/ibooking/repository/CampusRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.spm.ibooking.entity.*;
 
 @Repository
-public interface CampusRepository extends JpaRepository<Campus, Long> {
+public interface CampusRepository extends JpaRepository<Campus, Integer> {
 }
